# Optimize delta decoding (part 3)