--- a/encoding/delta/binary_packed_encode_amd64.s
+++ b/encoding/delta/binary_packed_encode_amd64.s
@@ -177,15 +177,15 @@ loop:
     VZEROUPPER
     RET
 
-// miniBlockPackInt32Default is the generic implementation of the algorithm to
+// encodeMiniBlockInt32Default is the generic implementation of the algorithm to
 // pack 32 bit integers into values of a given bit width (<=32).
 //
 // This algorithm is much slower than the vectorized versions, but is useful
 // as a reference implementation to run the tests against, and as fallback when
 // the code runs on a CPU which does not support the AVX2 instruction set.
 //
-// func miniBlockPackInt32Default(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
-TEXT ·miniBlockPackInt32Default(SB), NOSPLIT, $0-24
+// func encodeMiniBlockInt32Default(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
+TEXT ·encodeMiniBlockInt32Default(SB), NOSPLIT, $0-24
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
     MOVQ bitWidth+16(FP), R9
@@ -209,7 +209,7 @@ loop:
     JNE loop
     RET
 
-// miniBlockPackInt32x1bitAVX2 packs 32 bit integers into 1 bit values in the
+// encodeMiniBlockInt32x1bitAVX2 packs 32 bit integers into 1 bit values in the
 // the output buffer.
 //
 // The algorithm uses MOVMSKPS to extract the 8 relevant bits from the 8 values
@@ -218,8 +218,8 @@ loop:
 // block has 32 values (the block size is 128 and there are 4 mini blocks per
 // block).
 //
-// func miniBlockPackInt32x1bitAVX2(dst *byte, src *[miniBlockSize]int32)
-TEXT ·miniBlockPackInt32x1bitAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt32x1bitAVX2(dst *byte, src *[miniBlockSize]int32)
+TEXT ·encodeMiniBlockInt32x1bitAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
 
@@ -249,17 +249,17 @@ TEXT ·miniBlockPackInt32x1bitAVX2(SB), NOSPLIT, $0-16
     VZEROUPPER
     RET
 
-// miniBlockPackInt32x2bitsAVX2 implements an algorithm for packing 32 bit
+// encodeMiniBlockInt32x2bitsAVX2 implements an algorithm for packing 32 bit
 // integers into 2 bit values.
 //
-// The algorithm is derived from the one employed in miniBlockPackInt32x1bitAVX2
+// The algorithm is derived from the one employed in encodeMiniBlockInt32x1bitAVX2
 // but needs to perform a bit extra work since MOVMSKPS can only extract one bit
 // per packed integer of each YMM vector. We run two passes to extract the two
 // bits needed to compose each item of the result, and merge the values by
 // interleaving the first and second bits with PDEP.
 //
-// func miniBlockPackInt32x2bitsAVX2(dst *byte, src *[miniBlockSize]int32)
-TEXT ·miniBlockPackInt32x2bitsAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt32x2bitsAVX2(dst *byte, src *[miniBlockSize]int32)
+TEXT ·encodeMiniBlockInt32x2bitsAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
 
@@ -313,12 +313,12 @@ TEXT ·miniBlockPackInt32x2bitsAVX2(SB), NOSPLIT, $0-16
     VZEROUPPER
     RET
 
-// miniBlockPackInt32x32bitsAVX2 is a specialization of the bit packing logic
+// encodeMiniBlockInt32x32bitsAVX2 is a specialization of the bit packing logic
 // for 32 bit integers when the output bit width is also 32, in which case a
 // simple copy of the mini block to the output buffer produces the result.
 //
-// func miniBlockPackInt32x32bitsAVX2(dst *byte, src *[miniBlockSize]int32)
-TEXT ·miniBlockPackInt32x32bitsAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt32x32bitsAVX2(dst *byte, src *[miniBlockSize]int32)
+TEXT ·encodeMiniBlockInt32x32bitsAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
     VMOVDQU 0(BX), Y0
@@ -332,7 +332,7 @@ TEXT ·miniBlockPackInt32x32bitsAVX2(SB), NOSPLIT, $0-16
     VZEROUPPER
     RET
 
-// miniBlockPackInt32x3to16bitsAVX2 is the algorithm used to bit-pack 32 bit
+// encodeMiniBlockInt32x3to16bitsAVX2 is the algorithm used to bit-pack 32 bit
 // integers into values of width 3 to 16 bits.
 //
 // This function is a small overhead due to having to initialize registers with
@@ -359,7 +359,7 @@ TEXT ·miniBlockPackInt32x32bitsAVX2(SB), NOSPLIT, $0-16
 // which may not be aligned on byte boundaries so we shift the lower and upper
 // bits and compose two sets of 128 bits sequences (VPSLLQ, VPSRLQ, VBLENDPD),
 // merge them and write the 16 bytes result to the output buffer.
-TEXT ·miniBlockPackInt32x3to16bitsAVX2(SB), NOSPLIT, $0-24
+TEXT ·encodeMiniBlockInt32x3to16bitsAVX2(SB), NOSPLIT, $0-24
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
     MOVQ bitWidth+16(FP), CX
@@ -612,15 +612,15 @@ loop:
     VZEROUPPER
     RET
 
-// miniBlockPackInt64Default is the generic implementation of the algorithm to
+// encodeMiniBlockInt64Default is the generic implementation of the algorithm to
 // pack 64 bit integers into values of a given bit width (<=64).
 //
 // This algorithm is much slower than the vectorized versions, but is useful
 // as a reference implementation to run the tests against, and as fallback when
 // the code runs on a CPU which does not support the AVX2 instruction set.
 //
-// func miniBlockPackInt64Default(dst *byte, src *[miniBlockSize]int64, bitWidth uint)
-TEXT ·miniBlockPackInt64Default(SB), NOSPLIT, $0-24
+// func encodeMiniBlockInt64Default(dst *byte, src *[miniBlockSize]int64, bitWidth uint)
+TEXT ·encodeMiniBlockInt64Default(SB), NOSPLIT, $0-24
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
     MOVQ bitWidth+16(FP), R10
@@ -653,8 +653,8 @@ loop:
     JNE loop
     RET
 
-// func miniBlockPackInt64x1bitAVX2(dst *byte, src *[miniBlockSize]int64)
-TEXT ·miniBlockPackInt64x1bitAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt64x1bitAVX2(dst *byte, src *[miniBlockSize]int64)
+TEXT ·encodeMiniBlockInt64x1bitAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
 
@@ -705,8 +705,8 @@ TEXT ·miniBlockPackInt64x1bitAVX2(SB), NOSPLIT, $0-16
     VZEROUPPER
     RET
 
-// func miniBlockPackInt64x2bitsAVX2(dst *byte, src *[miniBlockSize]int64)
-TEXT ·miniBlockPackInt64x2bitsAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt64x2bitsAVX2(dst *byte, src *[miniBlockSize]int64)
+TEXT ·encodeMiniBlockInt64x2bitsAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
 
@@ -797,8 +797,8 @@ TEXT ·miniBlockPackInt64x2bitsAVX2(SB), NOSPLIT, $0-16
     VZEROUPPER
     RET
 
-// func miniBlockPackInt64x64bitsAVX2(dst *byte, src *[miniBlockSize]int64)
-TEXT ·miniBlockPackInt64x64bitsAVX2(SB), NOSPLIT, $0-16
+// func encodeMiniBlockInt64x64bitsAVX2(dst *byte, src *[miniBlockSize]int64)
+TEXT ·encodeMiniBlockInt64x64bitsAVX2(SB), NOSPLIT, $0-16
     MOVQ dst+0(FP), AX
     MOVQ src+8(FP), BX
     VMOVDQU 0(BX), Y0