--- a/encoding/delta/byte_array_amd64.s
+++ b/encoding/delta/byte_array_amd64.s
@@ -1,0 +1,9 @@
+//go:build !purego
+
+#include "funcdata.h"
+#include "textflag.h"
+
+// func validatePrefixAndSuffixLengthValuesAVX2(prefix, suffix []int32, maxLength int) (totalPrefixLength, totalSuffixLength int, ok bool)
+TEXT ·validatePrefixAndSuffixLengthValuesAVX2(SB), NOSPLIT, $0-73
+
+    RET