… encoding/delta/binary_packed_amd64_test.go
… encoding/delta/binary_packed_test.go
… encoding/delta/byte_array_purego.go
… encoding/delta/delta_amd64.go
… encoding/delta/delta_amd64.s
… encoding/delta/length_byte_array.go
… encoding/delta/length_byte_array_amd64.go
… encoding/delta/length_byte_array_amd64.s
… encoding/delta/length_byte_array_purego.go
… encoding/delta/length_byte_array_test.go
… encoding/encoding_test.go
… encoding/plain/plain_amd64.s
… encoding/rle/rle.go
… encoding/rle/rle_amd64.go
… encoding/rle/rle_amd64.s
… encoding/rle/rle_purego.go
… internal/bitpack/bitpack.go
… internal/bitpack/masks_int32_amd64.s
… internal/bitpack/pack.go
… internal/bitpack/unpack.go
… internal/bitpack/unpack_int32_amd64.go
… internal/bitpack/unpack_int32_amd64.s
… internal/bitpack/unpack_int32_purego.go
… internal/bitpack/unpack_int64_amd64.go
… internal/bitpack/unpack_int64_amd64.s
… internal/bitpack/unpack_int64_purego.go
… internal/bitpack/unpack_test.go
… page.go