Following up to , this PR contributes to https://github.com/segmentio/parquet-go/issues/226 by optimizing the decoding of DELTA_LENGTH_BYTE_ARRAY.

This encoding is the default used for variable length byte arrays (e.g. Go's `string` type), which makes it a very critical path since a lot of data are modeled as strings.

The main difficulty in improving these code path came from the fact that we often deal with short string values, so calls to `runtime.memmove` (via `copy`) end up having a high overhead since there are no opportunities to use SIMD instructions for the copy when the input is just a few bytes long.

In parquet, we always decode full pages, so I used an opportunistic copy strategy to improve the copy throughput:
* if we ensure that we have some padding at the end of the input and output buffers, we can always read fixed size chunks without being concerned about overflowing beyond the buffers limits
* with this assumption, we can make a second assumption that short strings will always fit within a handful of bytes
* the algorithm can always read chunks of 16 bytes from the input, and often end after the first loop iteration since most strings will fit within the first
* for longer strings, we can perform copies in even larger chunks (32 bytes seemed to show good results)

This strategy appeared to be effective on our tests, the following snapshots show the impact on benchmarks comparing throughput with the previous versions of the code, on the function itself, on the full decoding of pages, and on the merging of files containing a BYTE_ARRAY column:


```
name                               old time/op    new time/op     delta
DecodeLengthByteArray/maxLen=0       4.15µs ± 0%     4.58µs ± 0%   +10.19%  (p=0.000 n=9+8)
DecodeLengthByteArray/maxLen=10      4.88µs ± 0%     0.71µs ± 0%   -85.36%  (p=0.000 n=9+9)
DecodeLengthByteArray/maxLen=20      5.47µs ±27%     0.95µs ± 0%   -82.59%  (p=0.000 n=10+8)
DecodeLengthByteArray/maxLen=100     6.05µs ± 9%     2.55µs ± 0%   -57.92%  (p=0.000 n=9+9)
DecodeLengthByteArray/maxLen=1000    17.8µs ± 1%     15.1µs ± 0%   -15.35%  (p=0.000 n=10+10)

name                               old speed      new speed       delta
DecodeLengthByteArray/maxLen=0      963MB/s ± 0%    874MB/s ± 0%    -9.25%  (p=0.000 n=9+8)
DecodeLengthByteArray/maxLen=10    1.99GB/s ± 0%  13.60GB/s ± 0%  +583.19%  (p=0.000 n=9+9)
DecodeLengthByteArray/maxLen=20    2.76GB/s ±22%  15.59GB/s ± 0%  +465.63%  (p=0.000 n=10+8)
DecodeLengthByteArray/maxLen=100   9.01GB/s ± 8%  21.38GB/s ± 0%  +137.34%  (p=0.000 n=9+9)
DecodeLengthByteArray/maxLen=1000  28.8GB/s ± 1%   34.1GB/s ± 0%   +18.12%  (p=0.000 n=10+10)
```

```
name                                       old time/op    new time/op    delta
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array    76.5µs ± 1%    20.4µs ± 0%   -73.34%  (p=0.000 n=10+9)

name                                       old speed      new speed      delta
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array  1.89GB/s ± 1%  7.09GB/s ± 0%  +275.04%  (p=0.000 n=10+9)

name                                       old value/s    new value/s    delta
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array      131M ± 1%      491M ± 0%  +275.03%  (p=0.000 n=10+9)
```

```
name                                       old time/op  new time/op  delta
MergeFiles/BYTE_ARRAY/groups=3,rows=30000  25.3µs ± 2%  18.5µs ± 1%  -26.93%  (p=0.000 n=10+9)

name                                       old row/s    new row/s    delta
MergeFiles/BYTE_ARRAY/groups=3,rows=30000   38.9M ± 2%   53.2M ± 1%  +36.85%  (p=0.000 n=10+9)
```

I think this strategy should also be applicable to other decoding, but also encoding paths, which should yield great improvements to all reads and writes of string values.