@achille-roussel 2022-06-14
I applied the same decoding strategy to the DELTA_BYTE_ARRAY encoding which yielded similar results:

```
name                                             old time/op    new time/op    delta
Decode/DELTA_BYTE_ARRAY/byte_array                  151µs ± 3%      68µs ± 0%   -54.84%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array     104µs ± 1%      25µs ± 0%   -76.23%  (p=0.000 n=9+9)

name                                             old speed      new speed      delta
Decode/DELTA_BYTE_ARRAY/byte_array                960MB/s ± 3%  2187MB/s ± 0%  +127.78%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array  1.54GB/s ± 1%  6.47GB/s ± 0%  +320.63%  (p=0.000 n=9+9)

name                                             old value/s    new value/s    delta
Decode/DELTA_BYTE_ARRAY/byte_array                  66.4M ± 3%    147.1M ± 0%  +121.37%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array     96.2M ± 1%    404.6M ± 0%  +320.62%  (p=0.000 n=9+9)
```